Friday Final Push. Potential output fix.

+++ get_next_line.h
@@ -20,7 +20,7 @@ void	ft_strdel(void **as);
 char	*ft_strchr(const char *s, int c);
 char	*ft_substr(char const *s, unsigned int start, unsigned int len);
 char	*ft_strdup(const char *s1);
-size_t	ft_strlen(const char *s);
+unsigned int	ft_strlen(const char *s);
 void	*ft_memcpy(void *dst, const void *src, unsigned int n);
 char	*ft_strjoin(char const *s1, char const *s2);
 char	*get_next_line(int fd);
